Prefer generic provider partition settings over unset detected provider
Added a bit of defensiveness to merge s.t if a provider were to omit
these settings, we will still prefer the generic values.

This issue was specific to github as the github provider was relying on
implicitly unset partition nodes.

As a result, the generic provider's partition nodes were not preferred.