Ensure ability to run w/out partition when partition configured
We should be able to run the same suite using either the partition
command or the default run command. This determination is made by
whether or we've detected the command is intending to partition or not.

This was previously accounted for by checking to see if the partition
index and total were set to a value not equal to -1.

This was sufficient until a patch was introduced to fix how
CAPTAIN_PARTITION_* env variables were threaded into the partition
config.

That patch fixed the issue where the environment variables weren't
getting picked up, however, made the approach of determining whether or
not the command was running with partition insufficient.

This PR simplifies the determination of whether or not you are
partitioning to "is there a sensible partition total provided? (>= 1).

This handles both possible scenarios where the partition config is
defaulted to -1 or not set at all (ending up as a struct with a golang
"zero value").

In addition, an integration test is added to ensure that a test suite
with partitioning configured can in fact run without partitioning.