fix: escape the .net filter subsitution for shells
The filter substition is directly embedded into the command, which means
that `"` quotes will behave according to the shell. Resulting in
something along the lines of:

```sh
dotnet test --filter "FullyQualifiedName=Foo.Bar(val: "Hello World")"
```

The shell will interpret this and pass an invalid string to dotnet test,
resulting in the following error:

> 2023-08-01T20:20:59.7326742Z [xUnit.net 00:00:00.24] Cmg.DataLabGateway
> .IntegrationTests: Exception filtering tests: Incorrect format for
> TestCaseFilter Missing Operator '|' or '&'. Specify the correct format and
> try again. Note that the incorrect format can lead to no test getting
> executed.

This change will simply replace all occurrences of `"` with `\"`, which
should be interpreted by the shell correctly.