Run Build & Test CI weekly to avoid test file timings race.
When the tests don't run for a while, Cloud won't have the test file
timings that integration tests rely on. The first run will fail as it
relies on the manifest that Cloud now has, but has not yet processed
into timings.

By running this weekly, Cloud will always have a fresh-enough timing to
work with, and these tests should be stable and not require manual
intervention (deleting manifests).