Add endpoint compatibility based on host
This is required because our existing binary is tested against prior
versions of the integration tests.

These prior versions are setting the host name to captain.build -- in
which case we do not want to prefix the api endpoints with /captain
using the cloud host endpoint scheme.

Ideally I could check against cloud.rwx.com suffix, but that would break
local usage of cloud.local