Add omit prefix partition config
Playwright configures a "testDir" and reports test results relative to the test dir.

Captain will then record test results relative to this testDir.

When this happens, comparison with historical timing data will fail, falling back to round-robin.

This is because the server file ends up relative to the testDir, whereas captain's glob is the file on disk.

Supplying this option allows users to augment the client test file path for the sole purpose of matching.